Revised the usage of GTI and attitude files. Both file types must now have the FITS header keyword 'MJDREF', which defines the reference point of the entries in the 'START', 'STOP', and 'TIME' columns respectively. For handling GTI extension, now the corresponding routines from the library 'heautils' are applied instead of own code. The eROSITA raw event files the column time has been replaced by 'OTS' and 'FRACSEC'. For eROSITA event files in general MJDREF has been adjusted to the value 51544.0.

+++ tools/ero_calevents/ero_calevents.h
@@ -4,6 +4,7 @@
 
 #include "sixt.h"
 #include "attitude.h"
+#include "eroeventfile.h"
 #include "event.h"
 #include "eventfile.h"
 #include "gti.h"
@@ -38,16 +39,16 @@ struct Parameters {
 
 /** Event entry in an eROSITA calibrated event file. */
 typedef struct {
-  double time;
+  double time; /* [s] */
   long frame;
 
-  long pha;
-  float energy;
+  long pha; /* [adu] */
+  float energy; /* [eV] */
 
   int rawx, rawy;
   int subx, suby;
 
-  long ra, dec;
+  double ra, dec; /* [deg] */
   long x, y;
 
   long flag;
